Extract the PointerCapture helper classes
