+++ src/Uno.UI/UI/Xaml/UIElement.Pointers.wasm.cs
@@ -214,7 +214,7 @@ namespace Windows.UI.Xaml
 			}
 		}
 
-		partial void ReleasePointerCaptureNative(Pointer pointer)
+		partial void ReleasePointerNative(Pointer pointer)
 		{
 			var command = "Uno.UI.WindowManager.current.releasePointerCapture(" + HtmlId + ", " + pointer.PointerId + ");";
 			WebAssemblyRuntime.InvokeJS(command);
@@ -224,19 +224,6 @@ namespace Windows.UI.Xaml
 				SetStyle("touch-action", "auto");
 			}
 		}
-
-		//private void TryImplicitCapture(Pointer pointer)
-		//{
-		//	if (_allCaptures.TryGetValue(pointer, out var capture))
-		//	{
-		//		capture.AdditionalTargets.Add(new PointerCapture.CaptureTarget(this, isExplicit: false) {IsInNativeBubblingTree = true});
-		//	}
-		//	else
-		//	{
-				
-		//	}
-		//}
-
 		#endregion
 
 		#region HitTestVisibility
